+++ src/js/electron/ops/close-window.ts
@@ -1,4 +1,4 @@
-import {createOperation} from "../main-op"
+import {createOperation} from "../operations"
 
 export const closeWindow = createOperation("closeWindow", (main) => {
   main.windows.closeWindow()
